refactor polygonDrawing directory
