+++ src/client/map/GeometryDrawing/_layers/geometryLayer.ts
