+++ src/client/map/GeometryDrawing/_logic/onGeometryClick.ts
