+++ src/client/map/GeometryDrawing/_logic/onGeometryDrag.ts
