+++ src/client/map/GeometryDrawing/_logic/onGeometryHover.ts
