+++ src/client/map/GeometryDrawing/_types/geometryDrawingTypes.ts
